tidied up delete code

+++ service/routes.py
@@ -100,9 +100,11 @@ def remove_account(account_id):
     """
     app.logger.info("Request to delete an Account with id: %s", account_id)
     account = Account.find(account_id)
+    if not account:
+        abort(status.HTTP_404_NOT_FOUND, f"Account with id [{account_id}] could not be found.")
+
     account.delete()
-    # if not account:
-    #     abort(status.HTTP_404_NOT_FOUND, f"Account with id [{account_id}] could not be found.")
+
     return account.serialize(), status.HTTP_200_OK
